Fix toaster styling in petrinaut website
